pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "paper-digest-obsidian"
version = "0.1.0"
description = "Local paper summarization workflow for Codex and Obsidian."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
  { name = "thetan" }
]
keywords = ["arxiv", "obsidian", "paper-reading", "research", "summarization", "typer"]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Text Processing :: Markup :: Markdown",
  "Topic :: Utilities",
]
dependencies = [
  "beautifulsoup4>=4.13.0",
  "httpx>=0.28.0",
  "jinja2>=3.1.6",
  "lxml>=5.3.0",
  "pdfplumber>=0.11.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.7.0",
  "pymupdf>=1.25.0",
  "PyYAML>=6.0.2",
  "typer>=0.16.0",
]

[project.optional-dependencies]
dev = [
  "black>=25.1.0",
  "build>=1.2.2",
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.11.0",
]

[project.scripts]
paper-digest = "paper_digest.cli.app:main"
summarize-paper = "paper_digest.cli.app:run_summarize_paper"
summarize-topic = "paper_digest.cli.app:run_summarize_topic"
recommend-daily = "paper_digest.cli.app:run_recommend_daily"
start-my-day = "paper_digest.cli.app:run_recommend_daily"
search-notes = "paper_digest.cli.app:run_search_notes"
extract-images = "paper_digest.cli.app:run_extract_images"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "C4"]
ignore = ["B008"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--basetemp=test-output/pytest_tmp"
cache_dir = "test-output/pytest_cache"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_digest"]

[tool.hatch.build]
include = [
  "src/paper_digest/templates/*.j2",
  "examples/research_interests.yaml",
  "README.md",
  "LICENSE",
]
